SLING-12683: extract alias tests from MapEntriesTest
